requirements.txt
tk
dotenv 
psycopg2
bcrypt 
PILLOW